Rework apps logic + Implement Save + ble_title is now stateful
We don't need to connect to the devices in this step if you look to the future of the app.

we only need to know which of these devices are the cars, store them and go to the main page where the app will always start, and then connect to these devices.