Add Specific Search
This could stabilize the Rssi values even more

If there is only one device added to the final search, and that device has service Uuids then it will only scan for this device, else it will scan for all devices

I tried to implement this to search for all the devices in final device list only, but I faced a lot of problems and reverted to this, thankfully in the project we have one device only.

also the app will crash if there are any devices saved from previous builds.